Score screen, flow improvement, image buttons background

+++ src/com/mlaskows/quiz/model/entities/Level.java
@@ -117,7 +117,7 @@ public class Level {
 				pec++;
 			}
 		}
-		return ((int) (pec / ec)) * 100;
+		return (int) (((double) pec / (double) ec) * 100);
 	}
 
 	@Override
